Bugfix in precision calculation.

+++ src/geometry/common.rs
@@ -49,5 +49,5 @@ pub fn shader_texture(pos: &ProgramSource) -> ProgramSource {
     )
 }
 
-pub static PR_DEF : Precision = Precision::Float(23,16);
+pub static PR_DEF : Precision = Precision::Float(25,16);
 pub static PR_LOW : Precision = Precision::Float(5,8);
